Bugfix to ResultReporter
ResultReporter.reportResults is called on every experiment that is
queued in an assistant. This would result in an HTTP request being
attempted for each experiment, even though the assistant is told not to
report results anywhere. As a result, queuing a large number of
experiments would take a very long time. The fixed version does not
attempt an HTTP request if its report URL is set to null.