Start of conversion to Programmatic Spreadsheet
A first big commit where all table and plot manipulation is shifted from
objects of the MTNP library to objects of the new Programmatic
Spreadsheet library.

This version of the code still has about 100 compilation errors, all
related to handling of provenance which must be converted to objects of
the latest version of the Petit Poucet library. It is only intended as a
backup checkpoint.