Start of coding on version 3
Version 3 is a (mostly) complete rewrite of LabPal. The high-level
concepts (lab, experiment, assistant, tables, plots) remain the same,
but most of them are re-implemented and have their API changed in subtle
or drastic ways. The architecture of the whole platform is also modified
and, in most cases, simplified with respect to the previous version.

A significant difference in version 3 is the inclusion of unit tests for
most of the classes, a feature that was dearly missing in the version 2
codebase and which will hopefully result in more robust code.

This first commit corresponds to about a week worth of development (25
hours) done on a local machine before saving progress to this new branch.